Refactor firmware flashing support for improved integration
- Removed separate axiom_i2c_fw directory and merged all relevant code into the root directory to avoid redundancy and simplify diffs.
- Eliminated separate Makefile; integrated firmware build into the root-level Makefile. Module can now be built directly using the existing build system.
- Added kernel version checks wherever required for compatibility.
- Removed unrelated or incomplete additions unrelated to firmware flashing.
- Ensured minimal changes to existing files to improve maintainability and upstream compatibility.
